Modularize worldservice FastAPI layer
