+++ qmtl/worldservice/routers/__init__.py
@@ -1,0 +1,13 @@
+from .activation import create_activation_router
+from .bindings import create_bindings_router
+from .policies import create_policies_router
+from .validations import create_validations_router
+from .worlds import create_worlds_router
+
+__all__ = [
+    'create_activation_router',
+    'create_bindings_router',
+    'create_policies_router',
+    'create_validations_router',
+    'create_worlds_router',
+]
